fix double click camera cards
